+++ src/main/java/com/fintechiq/webzaim/controller/RequestContentController.java
@@ -1,7 +1,8 @@
-package com.fintechiq.Web_Zaim.controller;
+package com.fintechiq.webzaim.controller;
 
-import com.fintechiq.Web_Zaim.service.RequestContentService;
-import com.fintechiq.Web_Zaim.service.StopFactorCalculator;
+import com.fintechiq.webzaim.service.RequestContentService;
+import com.fintechiq.webzaim.service.StopFactorCalculator;
+import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,8 +10,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.io.IOException;
 
 @RestController
 @RequestMapping("/api/request-content")
@@ -30,8 +29,9 @@ public class RequestContentController {
         var requestContent = requestContentService.getRequestContent(id);
         return requestContent != null ? requestContent.getContent() : null;
     }
+
     @GetMapping("/calculator")
-    public boolean calculatorStopFactor( String regPerson, String verifiedName){
+    public boolean calculatorStopFactor(String regPerson, String verifiedName) {
         return stopFactorCalculator.calculatorStopFactor(regPerson, verifiedName);
     }
 }
