+++ src/main/java/com/fintechiq/webzaim/entity/AccountInfo.java
@@ -1,4 +1,4 @@
-package com.fintechiq.Web_Zaim.entity;
+package com.fintechiq.webzaim.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,9 +8,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import lombok.Data;
-
 import java.time.LocalDate;
+import lombok.Data;
 
 @Entity
 @Table(name = "account_info")
