+++ src/main/java/com/fintechiq/webzaim/entity/RegPerson.java
@@ -1,4 +1,4 @@
-package com.fintechiq.Web_Zaim.entity;
+package com.fintechiq.webzaim.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,17 +6,22 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import java.util.UUID;
 import lombok.Data;
 
-import java.util.UUID;
-
 @Entity
-@Table(name = "credit_bureau")
+@Table(name = "reg_person")
 @Data
-public class CreditBureau {
+public class RegPerson {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(name = "loan_request_id", nullable = false)
     private UUID loanRequestId;
+    @Column(name = "first_name")
+    private String firstname;
+    @Column(name = "middle_name")
+    private String middlename;
+    @Column(name = "last_name", nullable = false)
+    private String lastname;
 }
