+++ src/main/java/com/fintechiq/webzaim/entity/RequestContent.java
@@ -1,4 +1,4 @@
-package com.fintechiq.Web_Zaim.entity;
+package com.fintechiq.webzaim.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,9 +6,8 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import lombok.Data;
-
 import java.util.UUID;
+import lombok.Data;
 
 @Entity
 @Table(name = "request_content")
