+++ src/main/java/com/fintechiq/webzaim/repository/RegPersonRepository.java
@@ -1,6 +1,6 @@
-package com.fintechiq.Web_Zaim.repository;
+package com.fintechiq.webzaim.repository;
 
-import com.fintechiq.Web_Zaim.entity.RegPerson;
+import com.fintechiq.webzaim.entity.RegPerson;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
