+++ src/main/java/com/fintechiq/webzaim/repository/VerifiedNameRepository.java
@@ -1,6 +1,6 @@
-package com.fintechiq.Web_Zaim.repository;
+package com.fintechiq.webzaim.repository;
 
-import com.fintechiq.Web_Zaim.entity.VerifiedName;
+import com.fintechiq.webzaim.entity.VerifiedName;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
