--- src/test/java/com/fintechiq/Web_Zaim/WebZaimApplicationTests.java
@@ -1,12 +1,0 @@
-package com.fintechiq.Web_Zaim;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WebZaimApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-}
