+++ src/test/java/com/fintechiq/webzaim/WebZaimApplicationTests.java
@@ -1,0 +1,13 @@
+package com.fintechiq.webzaim;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WebZaimApplicationTests {
+
+	@Test
+	void contextLoads() {
+		// Этот метод оставлен пустым, так как он используется для проверки контекста загрузки Spring Boot
+	}
+}
